requirements.txt
lm-dataformat
tqdm
spacy
https://github.com/explosion/spacy-models/releases/download/pl_core_news_md-3.2.0/pl_core_news_md-3.2.0.tar.gz
requests
boto3
warcio
goose3
newspaper3k